Add fallback strategy for first time visits if not starting at home.
