+++ src/js/router/index.js
@@ -1,3 +1,4 @@
 export { HOME_PATH } from './constants';
 export { initRouter } from './init-router';
+export { isKnownRoutePath } from './route-resolution';
 export { renderRoute } from './render-route';
